refactor(1056): move ProjectActiviesView in buildProject feature
