+++ src/features/student/buildProject/views/ProjectActivitiesView/ProjectActivitiesView.test.ts
@@ -1,8 +1,8 @@
 import type { VueWrapper } from '@vue/test-utils'
 import { PageTitleStub } from '@/common/components/PageTitle/PageTitle.stub'
 import { ROUTES } from '@/common/constants'
-import { AllActivitiesTabStub } from '@/features/student/global/views/ProjectActivitiesView/components/AllActivitiesTab/AllActivitiesTab.stub'
-import ProjectActivitiesView from '@/features/student/global/views/ProjectActivitiesView/ProjectActivitiesView.vue'
+import { AllActivitiesTabStub } from '@/features/student/buildProject/views/ProjectActivitiesView/components/AllActivitiesTab/AllActivitiesTab.stub'
+import ProjectActivitiesView from '@/features/student/buildProject/views/ProjectActivitiesView/ProjectActivitiesView.vue'
 import { BddTest } from '@avenirs-esr/avenirs-dsav/test-utils'
 import { mountComponent } from 'tests/utils'
 import { beforeEach, expect, vi } from 'vitest'
