+++ src/features/student/global/locales/en.json
@@ -64,15 +64,6 @@
         }
       },
       "views": {
-        "projectActivitiesView": {
-          "allActivities": {
-            "header": {
-              "description": "Activities offered by your SUIO and your institution. For any requests for additional activities, please contact your guidance counselor.",
-              "seeAll": "See all activities"
-            }
-          },
-          "title": "My activities"
-        },
         "studentDeliverablesView": {
           "title": "(placeholder) All my deliverables"
         },
