+++ src/features/student/global/locales/fr.json
@@ -64,15 +64,6 @@
         }
       },
       "views": {
-        "projectActivitiesView": {
-          "allActivities": {
-            "header": {
-              "description": "Activités proposées par votre SUIO et votre établissement. Pour toute demande d’activités complémentaires, veuillez-vous rapprocher de votre conseiller d’orientation.",
-              "seeAll": "Voir toutes les activités"
-            }
-          },
-          "title": "Mes activités"
-        },
         "studentDeliverablesView": {
           "title": "(placeholder) Tous mes rendus"
         },
